requirements.txt
discord.py[voice]==1.2.5
ffmpeg-python
PyNaCl
opuslib
spotipy
